+++ packages/www/src/documentation/components/navigation.ts
@@ -111,6 +111,10 @@ const components: NavigationSection = {
           path: 'input-text',
         },
         {
+          title: 'InputTime',
+          path: 'input-time',
+        },
+        {
           title: 'InputTimeSelect',
           path: 'input-time-select',
         },